requirements.txt
requests==2.31.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
schedule==1.2.0
numpy==1.26.2
flask==3.0.0
